+++ lib/widgets/task_list_item.dart
@@ -32,6 +32,7 @@ class ReorderableTaskItem extends StatelessWidget {
   final VoidCallback onTap;
   final ValueChanged<bool?> onComplete;
   final VoidCallback? onDelete;
+  final VoidCallback? onSetDueToday;
   final ValueChanged<String>? onEditField;
 
   const ReorderableTaskItem({
@@ -44,6 +45,7 @@ class ReorderableTaskItem extends StatelessWidget {
     required this.onTap,
     required this.onComplete,
     this.onDelete,
+    this.onSetDueToday,
     this.onEditField,
   });
 
@@ -81,6 +83,7 @@ class ReorderableTaskItem extends StatelessWidget {
                   onTap: onTap,
                   onComplete: onComplete,
                   onEditField: onEditField,
+                  onSetDueToday: onSetDueToday,
                   onDelete: onDelete == null
                       ? null
                       : () async {